Clarify size range guard
Agent-Logs-Url: https://github.com/zezoallord2/Parallel-Computing/sessions/58ab2a9f-2611-4db9-81d5-add0c4a0c1a1

+++ src/main.cpp
@@ -68,7 +68,7 @@ size_t parse_size(const string& value, const string& flag) {
         if (consumed != value.size()) {
             fail("Invalid numeric value for " + flag + ": " + value);
         }
-        if (parsed > numeric_limits<size_t>::max()) {
+        if (parsed > static_cast<unsigned long long>(numeric_limits<size_t>::max())) {
             fail("Numeric value for " + flag + " is too large: " + value);
         }
         return static_cast<size_t>(parsed);
